Add bid service, implement gRPC

+++ src/SearchService/Consumers/AuctionFinishedConsumer.cs
@@ -14,7 +14,7 @@ namespace SearchService.Consumers
             if (context.Message.ItemSold)
             {
                 auction.Winner = context.Message.Winner;
-                auction.SoldAmount = context.Message.Amount;
+                auction.SoldAmount = (int)context.Message.Amount;
             }
 
             auction.Status = "Finished";
